Split out Literal into its own .h/.cpp. Also factor out common wasm-type module
