+++ src/wasm/wasm-type.cpp
@@ -1,0 +1,70 @@
+/*
+ * Copyright 2017 WebAssembly Community Group participants
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "wasm-type.h"
+
+#include <cstdlib>
+#include "compiler-support.h"
+
+namespace wasm {
+
+const char* printWasmType(WasmType type) {
+  switch (type) {
+    case WasmType::none: return "none";
+    case WasmType::i32: return "i32";
+    case WasmType::i64: return "i64";
+    case WasmType::f32: return "f32";
+    case WasmType::f64: return "f64";
+    case WasmType::unreachable: return "unreachable";
+    default: WASM_UNREACHABLE();
+  }
+}
+
+unsigned getWasmTypeSize(WasmType type) {
+  switch (type) {
+    case WasmType::none: abort();
+    case WasmType::i32: return 4;
+    case WasmType::i64: return 8;
+    case WasmType::f32: return 4;
+    case WasmType::f64: return 8;
+    default: WASM_UNREACHABLE();
+  }
+}
+
+bool isWasmTypeFloat(WasmType type) {
+  switch (type) {
+    case f32:
+    case f64: return true;
+    default: return false;
+  }
+}
+
+WasmType getWasmType(unsigned size, bool float_) {
+  if (size < 4) return WasmType::i32;
+  if (size == 4) return float_ ? WasmType::f32 : WasmType::i32;
+  if (size == 8) return float_ ? WasmType::f64 : WasmType::i64;
+  abort();
+}
+
+WasmType getReachableWasmType(WasmType a, WasmType b) {
+  return a != unreachable ? a : b;
+}
+
+bool isConcreteWasmType(WasmType type) {
+  return type != none && type != unreachable;
+}
+
+} // namespace wasm
